# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Image Comparison using Cosine Similarity Method",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "local host"
        },
        {
            "url": "http://localhost/image-comparison/api",
            "description": "in port 80"
        },
        {
            "url": "http://localhost/api",
            "description": "in port 80"
        },
        {
            "url": "http://20.20.20.42/image-comparison/api",
            "description": "TIF Akreditasi 2"
        }
    ],
    "paths": {
        "/compare": {
            "post": {
                "tags": [
                    "compare"
                ],
                "summary": "Compare two images, upload both.",
                "operationId": "Compare.compare",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "image1": {
                                        "description": "First image",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "image2": {
                                        "description": "Second image",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Invalid input"
                    },
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/lbph/test": {
            "get": {
                "tags": [
                    "face recognition"
                ],
                "summary": "Predict LBPH Face Recognition",
                "description": "Predict the LBPH Face Recognition model with an image.",
                "operationId": "testLBPHFaceRecognition",
                "responses": {
                    "200": {
                        "description": "Success response",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/lbph/train": {
            "post": {
                "tags": [
                    "face recognition"
                ],
                "summary": "Train LBPH Face Recognition",
                "description": "Train the LBPH Face Recognition model with an image.",
                "operationId": "trainLBPHFaceRecognition",
                "requestBody": {
                    "description": "Image data as form-data",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "image": {
                                        "description": "Image file (PNG or JPG format, less than 2048KB)",
                                        "type": "file"
                                    },
                                    "username": {
                                        "description": "username",
                                        "type": "text"
                                    },
                                    "replace": {
                                        "description": "replace",
                                        "type": "boolean",
                                        "default": "false"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Image trained"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Failed response - Validator failed from Laravel",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validator failed from Laravel"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Exception occurred"
                    }
                }
            }
        },
        "/lbph/predict": {
            "post": {
                "tags": [
                    "face recognition"
                ],
                "summary": "Predict LBPH Face Recognition",
                "description": "Predict the LBPH Face Recognition model with an image.",
                "operationId": "predictLBPHFaceRecognition",
                "requestBody": {
                    "description": "Image data as form-data",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "image": {
                                        "description": "Image file (PNG or JPG format, less than 2048KB)",
                                        "type": "file"
                                    },
                                    "username": {
                                        "description": "username",
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Image trained"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Failed response - Validator failed from Laravel",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validator failed from Laravel"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Exception occurred"
                    }
                }
            }
        },
        "/lbph/score/{filename}": {
            "get": {
                "tags": [
                    "scores"
                ],
                "summary": "Get score image",
                "description": "Retrieve a score image by filename.",
                "operationId": "3e0e30e6cad1105407da8ac874e2a64b",
                "parameters": [
                    {
                        "name": "filename",
                        "in": "path",
                        "description": "The filename of the image",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "image/jpeg": {}
                        }
                    },
                    "404": {
                        "description": "Image not found"
                    }
                }
            }
        },
        "/lbph/logs": {
            "get": {
                "tags": [
                    "logs"
                ],
                "summary": "Get logs",
                "description": "Retrieve a list of log entries",
                "operationId": "3761c17dfdcbbfb0198ae85a34c01c76",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "number"
                                            },
                                            "result_path": {
                                                "type": "string"
                                            },
                                            "user_id": {
                                                "type": "string"
                                            },
                                            "score": {
                                                "type": "number",
                                                "format": "decimal"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/lbph/reset": {
            "get": {
                "tags": [
                    "reset"
                ],
                "summary": "Reset LBPH Face Recognition",
                "description": "Reset the LBPH Face Recognition model.",
                "operationId": "b93da8eded1dce3c9bab6aef4a9ba691",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "reset success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}